Fixed incorrect comment and added more information
It turns out that rand only needs to be seeded once and then every subsequent
use of rand uses the previous number as a seed. So my reasoning that I shouldn't
use rand because it seeds to time every time is incorrect. Only the first number
is seeded to time. This would be a problem if someone ran the file several times
in the same second, since then they would get the same results. But this is
quite unlikely.

More importantly, there are other reasons to use <random> instead of rand, which
I explain in the comment.