Removed unnecessary command line argument
I realised it was unnecessary for there to be two command line arguments in the
c++ code, one for whether to randomise and one for which energy level. Instead,
we can use just one argument, for the energy level. This is because if the
energy level is greater than 1, we will definitely also need to randomise.